Remove install, contributor guides
Remove code to build singlehtml, since it was decided that it wasn't
really that useful, especially since it was basically just collection of
single-page docs rather than one doc.

Remove the install and contributor guides.  Due to the interconnected
nature of the documentation, this is tricky to do completely with
scripts.  Even with this commit, there are still more than 100
warnings generated about broken cross-document references (either
"undefined label"s or "unknown document"s).  It will require manual
effort to create a suitable text modification for each of these, and
then additional time to modify the script to programatically re-create
each one.